Complete code Refactoring - Verified and Tested
