+++ include/Functions.h
@@ -1,0 +1,18 @@
+// Functions from utils.cpp
+
+void println(String text, int column, int row, int text_size);
+void println(tm timeinfo, char *text, int column, int row, int text_size);
+
+// Functions from alarms.cpp
+void ring_alarm();
+void set_alarm(int alarm);
+
+// Functions from utils.cpp
+int wait_for_button_press();
+void go_to_menu();
+
+// FUnctions from temperature.cpp
+void check_temp();
+
+// Functions from time.cpp
+void set_time_zone();
